pyproject.toml
[tool.poetry]
name = "reviser"
version = "0.1.4"
description = "AWS Lambda function/layer version deployment manager."
authors = ["Scott Ernst <swernst@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://gitlab.com/rocket-boosters/reviser"
repository = "https://gitlab.com/rocket-boosters/reviser"
documentation = "https://gitlab.com/rocket-boosters/reviser"
keywords = ["aws", "lambda"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.8",
  "Topic :: Utilities",
  "Typing :: Typed"
]
exclude = ["reviser/tests"]

[tool.poetry.scripts]
reviser = 'reviser.cli:main'

[tool.poetry.dependencies]
python = "^3.8"
boto3 = { version = "^1.16.11", optional = true }
colorama = { version = "^0.4.4", optional = true }
Jinja2 = { version = "^2.11.2", optional = true }
prompt-toolkit = { version = "^3.0.8", optional = true }
PyYAML = { version = "^5.3.1", optional = true }
pipper = { version = "^0.7.0", optional = true }
toml = { version = "^0.10.2", optional = true }
poetry = { version = "^1.1.4", optional = true }

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
black = "yes"
coverage = "^5.3"
pytest-cov = "yes"
lobotomy = "^0.1.7"
flake8 = "^3.8.4"
flake8-black = "^0.2.1"
mypy = "^0.790"
taskipy = "^1.4.0"

[tool.taskipy.tasks]
# https://github.com/illBeRoy/taskipy
black = "black ."
mypy = "poetry run mypy ./reviser/ --ignore-missing-imports"
flake8 = "flake8 ."
test = "pytest . --cov-report=term-missing --cov=."
lint = "task black && task flake8 && task mypy"
check = "task black && task flake8 && task mypy && task test"
docs = "python docs_builder.py"
images = "python image_builder.py --publish"

[tool.poetry.extras]
shell = [
  "prompt-toolkit",
  "Jinja2",
  "colorama",
  "boto3",
  "PyYAML",
  "pipper",
  "poetry",
  "toml"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
